Refactor: rename MainUiManager folder to ui
